Refactor dev mode into separate file.
Auto-detect when we were installed in development mode at runtime.
